judgment-legal-agent/requirements.txt
anthropic>=0.39.0
pydantic>=2.0.0
google-genai>=1.0.0
judgeval
flask>=3.0.0
